elasticapm: split sender code to sender.go
